+++ src/main/java/com/noraknorak/user/presentation/dto/request/UserSignUpRequest.java
@@ -1,4 +1,4 @@
-package com.noraknorak.user.presentation.request;
+package com.noraknorak.user.presentation.dto.request;
 
 import com.noraknorak.user.domain.Role;
 import jakarta.validation.constraints.NotBlank;
@@ -7,7 +7,6 @@ import lombok.Data;
 
 @Data
 public class UserSignUpRequest {
-
     @NotBlank(message = "")
     private String name;
 
@@ -22,7 +21,4 @@ public class UserSignUpRequest {
 
     @NotNull(message = "")
     private Role role;
-
-    @NotBlank(message = "")
-    private String personalityTag;
 }
